+++ src/Circuit/Layer.hs
@@ -50,7 +50,7 @@ module Circuit.Layer
 where
 
 import Circuit.Category (Category (..))
-import Circuit.Channel (Assoc (..), Slide (..), Strength (..), Yank (..))
+import Circuit.Traced (Assoc (..), Slide (..), Strength (..), Yank (..))
 import Data.Kind (Constraint, Type)
 import Prelude hiding (id, (.))